refactor: use CamelCase (LB-9)
