+++ packages/client/src/components/ChangeAvatarDialog/index.ts
@@ -1,0 +1,1 @@
+export * from './ChangeAvatarDialog';
