+++ packages/client/src/components/ChangeAvatarForm/index.ts
@@ -1,0 +1,1 @@
+export * from './ChangeAvatarForm';
