+++ packages/client/src/components/ChangePasswordForm/index.ts
@@ -1,0 +1,1 @@
+export * from './ChangePasswordForm';
