+++ packages/client/src/components/SectionHeading/index.ts
@@ -1,0 +1,1 @@
+export * from './SectionHeading';
