+++ packages/client/src/components/UserInformationForm/index.ts
@@ -1,0 +1,1 @@
+export * from './UserInformationForm';
