--- packages/client/src/components/change-avatar-dialog/index.ts
@@ -1,1 +1,0 @@
-export * from './change-avatar-dialog';
