--- packages/client/src/components/change-avatar-form/index.ts
@@ -1,1 +1,0 @@
-export * from './change-avatar-form';
