--- packages/client/src/components/change-password-form/index.ts
@@ -1,1 +1,0 @@
-export * from './change-password-form';
