+++ packages/client/src/components/index.ts
@@ -1,6 +1,6 @@
-export * from './header';
-export * from './section-heading';
-export * from './user-information-form';
-export * from './change-password-form';
-export * from './change-avatar-dialog';
-export * from './change-avatar-form';
+export * from './Header';
+export * from './SectionHeading';
+export * from './UserInformationForm';
+export * from './ChangePasswordForm';
+export * from './ChangeAvatarDialog';
+export * from './ChangeAvatarForm';
