--- packages/client/src/components/section-heading/index.ts
@@ -1,1 +1,0 @@
-export * from './section-heading';
