--- packages/client/src/components/user-information-form/index.ts
@@ -1,1 +1,0 @@
-export * from './user-information-form';
