--- packages/client/src/pages/settings/index.ts
@@ -1,1 +1,0 @@
-export * from './settings';
